refactor(harnesstui): extract settings selection and surfaces
