+++ src/loushang/harnesstui/selection/__init__.py
@@ -1,0 +1,1 @@
+"""Shared terminal selection models."""
