+++ src/loushang/harnesstui/surface/__init__.py
@@ -1,0 +1,1 @@
+"""Shared terminal surface views."""
